+++ HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ---------------
 
+1.0.3 (2016-xx-xx)
+++++++++++++++++++
+* Added timeout to any TR64 action
+* Fix reboot action
+
 1.0.2 (2016-01-06)
 ++++++++++++++++++
 * Added the ability to request the namespace for a TR64 action
@@ -11,7 +16,7 @@ Release History
 * Improved error handling on XML parsing problems of any content
 * Changed the declaration of getGenericAssociatedDeviceInfo for consistency reasons
 * Changed for all Wifi actions that the interface id is now optional
-* Changed all TR64 action classes to a sub class of DeviceTR64
+* Changed all TR64 action classes to be a sub class of DeviceTR64
 * Fix an issue with relative URL's in the device definition
 
 1.0.1 (2016-01-04)